added tag
Now every mark has some tags and each item has a tag. A tag contains 2
part, one is the location title(will be displayed at most times) and the
other is the formatted address to indicate the  address(may be displayed
on a map). A tag also has a latitude and a longtitude attribute.